Add binary, re-org crate
