+++ eth2/proto_array_fork_choice/Cargo.toml
@@ -4,9 +4,16 @@ version = "0.1.0"
 authors = ["Paul Hauner <paul@sigmaprime.io>"]
 edition = "2018"
 
+[[bin]]
+name = "proto_array_fork_choice"
+path = "src/bin.rs"
+
 [dependencies]
 parking_lot = "0.9.0"
 types = { path = "../types" }
 itertools = "0.8.1"
 eth2_ssz = "0.1.2"
 eth2_ssz_derive = "0.1.0"
+serde = "1.0.102"
+serde_derive = "1.0.102"
+serde_yaml = "0.8.11"
