+++ eth2/proto_array_fork_choice/no_votes.yaml
@@ -1,0 +1,239 @@
+---
+finalized_block_slot: 0
+justified_epoch: 1
+finalized_epoch: 1
+finalized_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+operations:
+  - FindHead:
+      justified_epoch: 1
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000000
+  - ProcessBlock:
+      slot: 0
+      root: 0x0000000000000000000000000000000000000000000000000000000000000002
+      parent_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      justified_epoch: 1
+      finalized_epoch: 1
+  - FindHead:
+      justified_epoch: 1
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000002
+  - ProcessBlock:
+      slot: 0
+      root: 0x0000000000000000000000000000000000000000000000000000000000000001
+      parent_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      justified_epoch: 1
+      finalized_epoch: 1
+  - FindHead:
+      justified_epoch: 1
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000002
+  - ProcessBlock:
+      slot: 0
+      root: 0x0000000000000000000000000000000000000000000000000000000000000003
+      parent_root: 0x0000000000000000000000000000000000000000000000000000000000000001
+      justified_epoch: 1
+      finalized_epoch: 1
+  - FindHead:
+      justified_epoch: 1
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000002
+  - ProcessBlock:
+      slot: 0
+      root: 0x0000000000000000000000000000000000000000000000000000000000000004
+      parent_root: 0x0000000000000000000000000000000000000000000000000000000000000002
+      justified_epoch: 1
+      finalized_epoch: 1
+  - FindHead:
+      justified_epoch: 1
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000004
+  - ProcessBlock:
+      slot: 0
+      root: 0x0000000000000000000000000000000000000000000000000000000000000005
+      parent_root: 0x0000000000000000000000000000000000000000000000000000000000000004
+      justified_epoch: 2
+      finalized_epoch: 1
+  - FindHead:
+      justified_epoch: 1
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000000
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000004
+  - InvalidFindHead:
+      justified_epoch: 1
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000005
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+  - FindHead:
+      justified_epoch: 2
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000005
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000005
+  - ProcessBlock:
+      slot: 0
+      root: 0x0000000000000000000000000000000000000000000000000000000000000006
+      parent_root: 0x0000000000000000000000000000000000000000000000000000000000000005
+      justified_epoch: 2
+      finalized_epoch: 1
+  - FindHead:
+      justified_epoch: 2
+      justified_root: 0x0000000000000000000000000000000000000000000000000000000000000005
+      finalized_epoch: 1
+      justified_state_balances:
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+        - 0
+      expected_head: 0x0000000000000000000000000000000000000000000000000000000000000006
