+++ eth2/proto_array_fork_choice/src/error.rs
@@ -1,0 +1,31 @@
+use types::{Epoch, Hash256};
+
+#[derive(Clone, PartialEq, Debug)]
+pub enum Error {
+    FinalizedNodeUnknown(Hash256),
+    JustifiedNodeUnknown(Hash256),
+    InvalidFinalizedRootChange,
+    InvalidNodeIndex(usize),
+    InvalidParentIndex(usize),
+    InvalidBestChildIndex(usize),
+    InvalidJustifiedIndex(usize),
+    InvalidBestDescendant(usize),
+    InvalidParentDelta(usize),
+    InvalidNodeDelta(usize),
+    DeltaOverflow(usize),
+    IndexOverflow(&'static str),
+    InvalidDeltaLen {
+        deltas: usize,
+        indices: usize,
+    },
+    RevertedFinalizedEpoch {
+        current_finalized_epoch: Epoch,
+        new_finalized_epoch: Epoch,
+    },
+    InvalidBestNode {
+        justified_epoch: Epoch,
+        finalized_epoch: Epoch,
+        node_justified_epoch: Epoch,
+        node_finalized_epoch: Epoch,
+    },
+}
